requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
arrow==1.2.2
asgiref==3.4.1
async-timeout==4.0.2
asynctest==0.13.0
atomicwrites==1.4.0
attrs==21.4.0
boto3==1.23.10
botocore==1.26.10
certifi==2022.5.18.1
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
config==0.5.1
Cython==0.29.30
dataclasses==0.8
decorator==5.1.1
Django==3.2.13
django-cors-headers==3.10.1
djangorestframework==3.13.1
filelock==3.4.1
flatbuffers==2.0
frozenlist==1.2.0
gitdb==4.0.9
GitPython==3.1.18
gluonnlp==0.10.0
graphviz==0.8.4
halo==0.0.31
idna==2.6
idna-ssl==1.1.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
iniconfig==1.1.1
jmespath==0.10.0
joblib==1.1.0
koalanlp==2.1.7
kobert @ git+https://git@github.com/SKTBrain/KoBERT.git@e1f2f37055e7460d8427f6912579c0162cb69831
kss==2.5.1
log-symbols==0.0.14
multidict==5.2.0
mxnet==1.7.0.post2
ngrok==0.0.1
numpy==1.19.5
onnxruntime==1.8.0
packaging==21.3
pandas==1.1.5
pluggy==1.0.0
protobuf==3.19.4
py==1.11.0
py4j==0.10.9.5
pyparsing==3.0.7
pytest==7.0.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
regex==2022.6.2
requests==2.27.1
requests-toolbelt==0.9.1
s3transfer==0.5.2
sacremoses==0.0.53
sentencepiece==0.1.96
six==1.16.0
smmap==5.0.0
spinners==0.0.24
sqlparse==0.4.2
tabulate==0.8.9
termcolor==1.1.0
tokenizers==0.8.1rc2
tomli==1.2.3
torch==1.10.2
tqdm==4.64.0
transform==1.0.21
transform-tools==1.0.3
transformers==3.1.0
typing_extensions==4.1.1
update-checker==0.18.0
urllib3==1.26.9
validators==0.19.0
whitenoise==5.3.0
yarl==1.7.2
zipp==3.6.0
